+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `axi_test`: The random AXI master `rand_axi_master` can now emit wrapping bursts (but does not do
   so by default).  Three new parameters control the burst types of the emitted transactions; not
   setting those parameters means the random master behaves as it did before this change.
+- Interface `AXI_BUS_DV`: Add `Monitor` modport, in which all signals are inputs.
+- `axi/assign.svh`: Add `AXI_ASSIGN_MONITOR` macro, which assigns an `AXI_BUS` to an
+  `AXI_BUS_DV.Monitor`.
+- Package `axi_test`: Add `axi_scoreboard` class, which checks that data read from a memory address
+  matches data written to that address.
 
 ### Changed
 - `axi_pkg:`